requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
cryptography
sqlalchemy
aiosqlite
alembic
python-multipart
google-genai
pytest
pytest-asyncio
